amazing_videodoc-main/requirements.txt
# 基础依赖
# 需要系统安装ffmpeg二进制工具

# Agent 框架依赖
agno>=0.1.0               # Agent 框架
lancedb>=0.3.0            # 向量数据库
pyarrow>=10.0.0           # LanceDB依赖

# 视频帧去重功能依赖
cohere>=5.0.0             # 用于调用Cohere API
numpy>=1.21.0             # 用于向量计算
Pillow>=8.0.0             # 用于图片处理
scipy>=1.7.0              # 用于高级数学计算

# 图文混排笔记功能依赖
markdown-pdf>=0.1.0       # 用于Markdown转PDF生成
markdown>=3.4.0           # 用于Markdown转HTML

cryptography>=41.0.0      # RSA签名与验签
fastapi
uvicorn[standard]
python-multipart
requests
yt-dlp
pydantic
python-dotenv
openai
prometheus-fastapi-instrumentator
sentry-sdk[fastapi]

faiss-cpu
pycryptodome>=3.20.0
